server/requirements.txt
fastapi
uvicorn
python-dotenv
llama-parse
pandas
python-multipart
pypdf
sqlalchemy
google-generativeai
